refactor(startup): move banner helpers out of entrypoint
